// apps/desktop/src/ipc/generated/PackageExchangeSources.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PackageDraftOption } from "./PackageDraftOption";
import type { PackageOption } from "./PackageOption";

export type PackageExchangeSources = { drafts: Array<PackageDraftOption>, acts: Array<PackageOption>, productionSource: PackageOption | null, schedule: PackageOption | null, callSheets: Array<PackageOption>, };